polish: fix expression-translator overflow + long-read seq/qual truncation
Bug fixes found in a full review + ASan pass:
- translate_ops allocated 2n+64 but a run of unary '!' expands 4x
  ("not "); a ~>32-! expression overflowed the heap buffer. Bump to 4n+64
  (verified under ASan with a 40-! expression).
- seq/sequence/qual and per-base 'base' were truncated at a fixed 1024-byte
  buffer, corrupting reads >1023 bp (e.g. long-read sequencing). The buffer
  is now sized to the read length (alloca, heap fallback for absurd reads);
  verified on a 2000 bp read: slen(seq)==2000, base@qpos1500 correct.

Cleanup: uint32 loop counters / flag comparisons (sign-compare), ruff
format. Regression tests added (long '!' chain, 2000 bp read slen(seq)).
ASan battery (mutation/base, tags/regex/per-base/-p, overflow case): clean.

+++ backend/countmut_core.c
@@ -100,8 +100,8 @@ static int base_to_index(char c) {
 
 /* read-level filters (samtools reqflags/exclflags + mapq + NS + bisulfite tags) */
 static int read_fails(const cm_config *cfg, const bam1_t *b) {
-    if (cfg->req_flags && (b->core.flag & cfg->req_flags) != cfg->req_flags) return 1;
-    if (cfg->excl_flags && (b->core.flag & cfg->excl_flags)) return 1;
+    if (cfg->req_flags && (b->core.flag & (uint32_t)cfg->req_flags) != (uint32_t)cfg->req_flags) return 1;
+    if (cfg->excl_flags && (b->core.flag & (uint32_t)cfg->excl_flags)) return 1;
     if ((int)b->core.qual < cfg->min_mapq) return 1;
     if (cfg->max_sub >= 0) {
         uint8_t *aux = bam_aux_get(b, "NS");
